+++ elex-testing-results/helpers/filterData.js
@@ -1,0 +1,11 @@
+async function filterData(testData) {
+  const yesterdaysDate = new Date(new Date().setDate(new Date().getDate() - 1));
+
+  return testData.filter(
+    (data) => new Date(`${data.testDate}T00:00`) > yesterdaysDate
+  );
+}
+
+module.exports = {
+  filterData,
+};
